requirements.txt
asgiref==3.8.1
Django==5.1.1
pillow==10.4.0
sqlparse==0.5.1
djangorestframework==3.14.0
twilio==8.4.0
tzdata==2024.2
